Extract TransactionManager interface
