+++ presto-main/src/test/java/com/facebook/presto/transaction/TestTransactionManager.java
@@ -70,7 +70,7 @@ public class TestTransactionManager
     {
         try (IdleCheckExecutor executor = new IdleCheckExecutor()) {
             CatalogManager catalogManager = new CatalogManager();
-            TransactionManager transactionManager = TransactionManager.create(new TransactionManagerConfig(), executor.getExecutor(), catalogManager, finishingExecutor);
+            TransactionManager transactionManager = InMemoryTransactionManager.create(new TransactionManagerConfig(), executor.getExecutor(), catalogManager, finishingExecutor);
 
             Connector c1 = new TpchConnectorFactory().create(CATALOG_NAME, ImmutableMap.of(), new TestingConnectorContext());
             registerConnector(catalogManager, transactionManager, CATALOG_NAME, CONNECTOR_ID, c1);
@@ -100,7 +100,7 @@ public class TestTransactionManager
     {
         try (IdleCheckExecutor executor = new IdleCheckExecutor()) {
             CatalogManager catalogManager = new CatalogManager();
-            TransactionManager transactionManager = TransactionManager.create(new TransactionManagerConfig(), executor.getExecutor(), catalogManager, finishingExecutor);
+            TransactionManager transactionManager = InMemoryTransactionManager.create(new TransactionManagerConfig(), executor.getExecutor(), catalogManager, finishingExecutor);
 
             Connector c1 = new TpchConnectorFactory().create(CATALOG_NAME, ImmutableMap.of(), new TestingConnectorContext());
             registerConnector(catalogManager, transactionManager, CATALOG_NAME, CONNECTOR_ID, c1);
@@ -130,7 +130,7 @@ public class TestTransactionManager
     {
         try (IdleCheckExecutor executor = new IdleCheckExecutor()) {
             CatalogManager catalogManager = new CatalogManager();
-            TransactionManager transactionManager = TransactionManager.create(new TransactionManagerConfig(), executor.getExecutor(), catalogManager, finishingExecutor);
+            TransactionManager transactionManager = InMemoryTransactionManager.create(new TransactionManagerConfig(), executor.getExecutor(), catalogManager, finishingExecutor);
 
             Connector c1 = new TpchConnectorFactory().create(CATALOG_NAME, ImmutableMap.of(), new TestingConnectorContext());
             registerConnector(catalogManager, transactionManager, CATALOG_NAME, CONNECTOR_ID, c1);
@@ -172,7 +172,7 @@ public class TestTransactionManager
             throws Exception
     {
         try (IdleCheckExecutor executor = new IdleCheckExecutor()) {
-            TransactionManager transactionManager = TransactionManager.create(
+            TransactionManager transactionManager = InMemoryTransactionManager.create(
                     new TransactionManagerConfig()
                             .setIdleTimeout(new Duration(1, TimeUnit.MILLISECONDS))
                             .setIdleCheckInterval(new Duration(5, TimeUnit.MILLISECONDS)),
